make RelayProxyTrait compatible with relay extension 0.9.0
