+++ src/Symfony/Component/Cache/Traits/Relay/CopyTrait.php
@@ -1,0 +1,36 @@
+<?php
+
+/*
+ * This file is part of the Symfony package.
+ *
+ * (c) Fabien Potencier <fabien@symfony.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
+namespace Symfony\Component\Cache\Traits\Relay;
+
+if (version_compare(phpversion('relay'), '0.8.1', '>=')) {
+    /**
+     * @internal
+     */
+    trait CopyTrait
+    {
+        public function copy($src, $dst, $options = null): \Relay\Relay|bool
+        {
+            return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->copy(...\func_get_args());
+        }
+    }
+} else {
+    /**
+     * @internal
+     */
+    trait CopyTrait
+    {
+        public function copy($src, $dst, $options = null): \Relay\Relay|false|int
+        {
+            return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->copy(...\func_get_args());
+        }
+    }
+}
