+++ src/Symfony/Component/Cache/Traits/RelayProxyTrait.php
@@ -17,11 +17,6 @@ if (version_compare(phpversion('relay'), '0.8.1', '>=')) {
      */
     trait RelayProxyTrait
     {
-        public function copy($src, $dst, $options = null): \Relay\Relay|bool
-        {
-            return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->copy(...\func_get_args());
-        }
-
         public function jsonArrAppend($key, $value_or_array, $path = null): \Relay\Relay|array|false
         {
             return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->jsonArrAppend(...\func_get_args());
@@ -148,9 +143,5 @@ if (version_compare(phpversion('relay'), '0.8.1', '>=')) {
      */
     trait RelayProxyTrait
     {
-        public function copy($src, $dst, $options = null): \Relay\Relay|false|int
-        {
-            return ($this->lazyObjectState->realInstance ??= ($this->lazyObjectState->initializer)())->copy(...\func_get_args());
-        }
     }
 }
